requirements.txt
Django
django-cors-headers
django-crispy-forms
django-dotenv
django-model-utils
djangorestframework
djangorestframework-simplejwt
gunicornDjango==4.1.6
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-filter==22.1
django-model-utils==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
